requirements.txt
pyTelegramBotAPI
telethon
pymongo
python-dotenv
motor  # For async MongoDB operations
flask
requests